+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 1.2.0
 
+* Added support for partial payments.
+* Added loading bottom sheet for the advanced flow google pay component.
 * Improved dynamic viewport of card component.
-* Added the missing loading bottom sheet for the advanced flow google pay component.
 * Updated iOS SDK to v5.13.0.
 
 ## 1.1.0