+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 * Improved dynamic viewport of card component.
 * Added the missing loading bottom sheet for the advanced flow google pay component.
-* Updated iOS SDK to v5.13.0.
+* Updated iOS SDK to v5.14.0.
 
 ## 1.1.0